+++ NEWS.md
@@ -1,3 +1,6 @@
+# audit.base 0.4.0 _2023-06-10_
+- refactor: Move code from audit.connect to audit.base for reuse
+
 # audit.base 0.3.0 _2023-06-05_
 - feat: Add GitHub Actions
 - fix: Use new package name